infrastructure/cdk/requirements.txt
aws-cdk-lib>=2.150.0,<3
constructs>=10.4.0,<11
